Resolved the reviews
